use ARMCI not MA mem in GA_Transpose
GA_Transpose uses ga_malloc, which defaults to using
MA memory.  this means that GA requires MA to be initialized.
this changes GA_Transpose to use ARMCI memory.

if ARMCI_Malloc_local is much slower than ga_malloc, this might
be noticeable (because the block is called in the loop)
but that seems unlikely at this point.

it is trivial to revert to the old implementation by undefining
the macro (GA_TRANSPOSE_USE_ARMCI_MEM) that enables this.

fix issue #163

Signed-off-by: Jeff Hammond <jeff.science@gmail.com>